% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_read.R
\name{coerce_path_segments}
\alias{coerce_path_segments}
\title{Safely coerce various path specs to segment vector
Accepts: list("a","b"), c("a","b"), "a.b", 'list("a","b")'}
\usage{
coerce_path_segments(p)
}
\arguments{
\item{p}{mixed}
}
\description{
Safely coerce various path specs to segment vector
Accepts: list("a","b"), c("a","b"), "a.b", 'list("a","b")'
}
\keyword{internal}
