% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_build_input.R
\name{complete_headers_for_pass2}
\alias{complete_headers_for_pass2}
\title{Complete headers minimally for Pass-2 (keeps user values)}
\usage{
complete_headers_for_pass2(pass1_headers, want_streaming = NA)
}
\arguments{
\item{pass1_headers}{list}

\item{want_streaming}{logical or NA}
}
\value{
list headers_p2
}
\description{
Complete headers minimally for Pass-2 (keeps user values)
}
