% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_io.R
\name{ensure_registry_header}
\alias{ensure_registry_header}
\title{Ensure registry directory exists and header is present}
\usage{
ensure_registry_header(path = get_registry_path())
}
\arguments{
\item{path}{character(1) registry yaml path}
}
\description{
Creates the Registry directory and writes a file header if the Registry
file does not exist yet or is empty.
}
