% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_utils.R
\name{escape_replacement}
\alias{escape_replacement}
\title{Escape replacement string for gsub(perl=TRUE) (escapes '\' and '$')}
\usage{
escape_replacement(x)
}
\description{
Escape replacement string for gsub(perl=TRUE) (escapes '\' and '$')
}
\keyword{internal}
