% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_utils.R
\name{flat_key_to_regex}
\alias{flat_key_to_regex}
\title{Turn a flat key with ".." into a regex for any numeric index
"choices..delta.content" -> "^choices\\.[0-9]+\\.delta\\.content$"
Handles start/middle/end ".." safely.}
\usage{
flat_key_to_regex(key)
}
\arguments{
\item{key}{character(1) flat key}
}
\value{
character(1) regex
}
\description{
Turn a flat key with ".." into a regex for any numeric index
"choices..delta.content" -> "^choices\\.[0-9]+\\.delta\\.content$"
Handles start/middle/end ".." safely.
}
\keyword{internal}
