% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_rank_endpoint.R
\name{hit_delta_flag}
\alias{hit_delta_flag}
\title{Strong-signal (delta-like) hit for streaming}
\usage{
hit_delta_flag(path, key, strong_tokens)
}
\arguments{
\item{path}{character}

\item{key}{character}

\item{strong_tokens}{character vector}
}
\value{
0/1 integer
}
\description{
Strong-signal (delta-like) hit for streaming
}
\keyword{internal}
