% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_build_input.R
\name{inject_optional_params}
\alias{inject_optional_params}
\title{Inject optional params into a body using a whitelist}
\usage{
inject_optional_params(body, optional_params = NULL, whitelist = NULL)
}
\arguments{
\item{body}{list}

\item{optional_params}{named list of candidates to inject}

\item{whitelist}{character vector of allowed keys}
}
\description{
Inject optional params into a body using a whitelist
}
