% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_utils.R
\name{is_scalar_chr}
\alias{is_scalar_chr}
\title{Return TRUE if x is a length-1, non-NA character}
\usage{
is_scalar_chr(x)
}
\description{
Return TRUE if x is a length-1, non-NA character
}
\keyword{internal}
