% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_utils.R
\name{normalize_path_key_with_regex}
\alias{normalize_path_key_with_regex}
\title{Normalize any path spec to a flat key AND, if needed, a regex with wildcard index}
\usage{
normalize_path_key_with_regex(path)
}
\arguments{
\item{path}{mixed (list("a","b"), "a.b", 'list("a","b")')}
}
\value{
list(key=character(1), regex=character(1) or NA)
}
\description{
Normalize any path spec to a flat key AND, if needed, a regex with wildcard index
}
