% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_rank_endpoint.R
\name{softmax_safe}
\alias{softmax_safe}
\title{Safe softmax}
\usage{
softmax_safe(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric probabilities summing to 1 (or zero vector if all -Inf-like)
}
\description{
Safe softmax
}
\keyword{internal}
