% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_utils.R
\name{unwrap_typed_defaults}
\alias{unwrap_typed_defaults}
\title{Unwrap typed optional_defaults (value/type) back to R types}
\usage{
unwrap_typed_defaults(lst)
}
\description{
Unwrap typed optional_defaults (value/type) back to R types
}
\keyword{internal}
