% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_utils.R
\name{wrap_typed_defaults}
\alias{wrap_typed_defaults}
\title{Wrap parameters with type info for registry}
\usage{
wrap_typed_defaults(lst)
}
\description{
Wrap parameters with type info for registry
}
\keyword{internal}
