% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar.R
\name{update_progress_bar}
\alias{update_progress_bar}
\title{Update progress bar for experiment}
\usage{
update_progress_bar(
  current,
  total,
  start_time,
  bar_width = 40,
  model_name = ""
)
}
\arguments{
\item{current}{当前完成的试次数}

\item{total}{总试次数}

\item{start_time}{实验开始时间}

\item{bar_width}{进度条宽度}

\item{model_name}{模型名，用于显示}
}
\value{
NULL (直接在控制台打印进度条)
}
\description{
Update progress bar for experiment
}
